pyproject.toml
[tool.poetry]
name = "42checkin"
version = "0.1.0"
description = ""
authors = ["silim124 <siniim0124@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flake8 = "^3.9.2"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"
Django = "^3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
